+++ pytsite/image/_functions.py
@@ -9,10 +9,11 @@ from pytsite import file as _file
 from . import _model
 
 
-def create(source_path: str, name: str=None, description: str=None) -> _model.Image:
+def create(source_path: str, name: str=None, description: str=None, remove_source: bool=False,
+           propose_store_path: str=None) -> _model.Image:
     """Create an image from URL or local file.
     """
-    img_entity = _file.manager.create(source_path, name, description, 'image')
+    img_entity = _file.functions.create(source_path, name, description, 'image', remove_source, propose_store_path)
 
     mime = str(img_entity.f_get('mime'))
     if not mime.endswith(('png', 'jpeg', 'gif')):
@@ -29,4 +30,4 @@ def create(source_path: str, name: str=None, description: str=None) -> _model.Im
 def get(uid: str=None, rel_path: str=None) -> _model.Image:
     """Get image.
     """
-    return _file.manager.get(uid, rel_path, 'image')
+    return _file.functions.get(uid, rel_path, 'image')